adding tags to tags, not yet to articles
