+++ pages/articles/create/index.js
@@ -1,15 +1,13 @@
 import NormEdit from '../../../components/markdown-editor/MarkdownEditor'
 import Create_Article from '../../api/create'
-//import checkTags from '../../api/tags'
-//import sendTags from '../../api/tags'
 //import getTagID from '../../api/tags'
 import * as React from 'react'
 import Styles from './create.module.css'
 import ReactTagInput from '@pathofdev/react-tag-input'
 import '@pathofdev/react-tag-input/build/index.css'
-import send from '../../api/tags_2'
+import send from '../../api/tags'
 
-export default function Index(/* { props }*/) {
+export default function Index({ props }) {
   const [value, setValue] = React.useState('**Hello world!!!**')
   const [title, setTitle] = React.useState('title')
   const [tags, setTags] = React.useState(['example tag'])
@@ -20,15 +18,24 @@ export default function Index(/* { props }*/) {
   }
   const handleSubmit = async (event) => {
     event.preventDefault()
-    /*
-    console.log(props.allTags.length)
-    const allTags = props.allTags
-    const o = sendTags(tags, allTags, checkTags)
-    console.log(o)
-    //console.log('sendTags: ' + sendTags(tags, allTags, checkTags))
-    */
 
-    send(tags)
+    var i
+    var x
+    var y
+
+    for (i = 0; i < tags.length; i++) {
+      y = 1
+      for (x = 0; x < props.allTags.length; x++) {
+        console.log('compairison: ' + tags[i] + '--' + props.allTags[x].TagName)
+        if (tags[i] == props.allTags[x].TagName) {
+          y = 0
+        }
+      }
+      if (y == 1) {
+        send(tags[i])
+      }
+    }
+
     const newArticleData = await Create_Article(title, value, tags)
     setNewArticle(newArticleData)
   }
